feat: signup form remove unused svg move form validation file to /validations/signup change form validation attrs
