+++ frontend/package.json
@@ -89,7 +89,6 @@
     "slugify": "^1.6.5",
     "tailwindcss": "^3.0.23",
     "use-debounce": "^6.0.1",
-    "validate.js": "^0.13.1",
     "yup": "^0.32.11"
   },
   "devDependencies": {
